Changed scan functions to return coordinates and measured values out. Moved break functions to look for overall deviation from a reference value in geometry_search, as this means that the function can be reused to look for defects (if liftoff can be corrected) Also changed break functions in trace_geometry to work out when the measured value is closer to one of two reference values, meaning that it can be found on or off the geometry. Changed names: geometry_search -> domain_search; trace_geometry -> trace_perimeter.
Need to debug the new break functions: may have to revert to old method if they don't work any more.
Need to debug to make sure we can find volumetric defects (i.e. anything that isn't a crack)
Need to process the geometry coordinates to return the approximate shape. Maybe make a geometry class and build in some default shapes to look for?